requirements.txt
# Core requirements
google-ads==18.0.0
google-auth==2.15.0
google-auth-oauthlib==0.7.1
google-auth-httplib2==0.1.0
googleads==33.0.0
protobuf==4.21.12
pyyaml==6.0

# Data processing
pandas==1.5.2
numpy==1.23.5

# Machine learning
scikit-learn==1.2.0
joblib==1.2.0

# Reporting and visualization
matplotlib==3.6.2
seaborn==0.12.2
plotly==5.11.0
tabulate==0.9.0

# HTTP and network
requests==2.28.1
aiohttp==3.8.3

# Utility libraries
python-dateutil==2.8.2
tqdm==4.64.1
pytz==2022.7

# Development and testing
pytest==7.2.0
pytest-cov==4.0.0
mypy==0.991
black==22.12.0
flake8==6.0.0

# Email and notification
sendgrid==6.9.7

# For Slack notifications
slack-sdk==3.19.5

# For Cloud deployment
google-cloud-storage==2.7.0 